trying approach with paper.js instead
